+++ pkg/detectors/clickuppersonaltoken/clickuppersonaltoken.go
@@ -2,9 +2,10 @@ package clickuppersonaltoken
 
 import (
 	"context"
-	regexp "github.com/wasilibs/go-re2"
 	"net/http"
 	"strings"
+
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
